backend/requirements/base.txt
coreapi==2.3.3
django-cors-headers==2.4.0
django-filter==2.0.0
Django==2.2.13
djangorestframework-camel-case==1.0b1
djangorestframework-jwt==1.11.0
djangorestframework==3.9.1
Markdown==3.0.1
Pygments==2.3.1
django-countries==5.3.3
django-import-export==2.0.2
serverless-wsgi==1.7.4
django-money==1.1
psycopg2-binary==2.8.5
